Compacted and completed Spawners
- Removed IronSpawner
- Removed CoalSpawner
+ Added BaseSpawner
+ Added Spawn locations
