+++ ConceptDevelopmentAssessment/Assembly-CSharp.csproj
@@ -62,7 +62,6 @@
     <Compile Include="Assets\Scripts\FragmentScript.cs" />
     <Compile Include="Assets\Scripts\PlayerScript.cs" />
     <Compile Include="Assets\Scripts\FadeScript.cs" />
-    <Compile Include="Assets\Scripts\IronSpawner.cs" />
     <Compile Include="Assets\Scripts\DrownScript.cs" />
     <Compile Include="Assets\Scripts\MainMenu.cs" />
     <Compile Include="Assets\Scripts\Camera.cs" />
@@ -73,11 +72,11 @@
     <Compile Include="Assets\Scripts\PlayerMovement.cs" />
     <Compile Include="Assets\Scripts\LanternCollector.cs" />
     <Compile Include="Assets\Scripts\IntroScript.cs" />
-    <Compile Include="Assets\Scripts\CoalSpawner.cs" />
     <Compile Include="Assets\Scripts\Water Level.cs" />
     <Compile Include="Assets\Scripts\GroundCheck.cs" />
     <Compile Include="Assets\Scripts\CamMov.cs" />
     <Compile Include="Assets\Scripts\GameOver.cs" />
+    <Compile Include="Assets\Scripts\BaseSpawner.cs" />
     <Compile Include="Assets\Scripts\CoalScript.cs" />
   </ItemGroup>
   <ItemGroup>
