+++ ConceptDevelopmentAssessment/Assets/Scripts/BaseSpawner.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 95656c3edb9fd6647b6008e5e19ab7f4
+guid: 922c2380962205d41958c2ade582d00e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
