--- ConceptDevelopmentAssessment/Assets/Scripts/IronSpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ef2c30b097c70e749b01fee7566ea57e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
